Add namespace to tests, small change to test for additional test case
Put tests into namespaces.

Modify json in test such that if fish -> location -> time parsing fails for arrays, the fish in the test named "Fish from A night or sunset" will return any time, such that it will match location at "StopA" in voyage "C by Night". This would result in an extra location for this fish, which should fail the test.